requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
pydantic==2.6.1
langchain==0.1.9
langchain-community==0.0.24
chromadb==0.4.22
sentence-transformers==2.5.1
transformers==4.37.2
torch==2.2.0
pypdf==4.0.1
beautifulsoup4==4.12.3
unstructured==0.12.3
python-dotenv==1.0.1
finnhub-python==2.4.19
yfinance==0.2.36
peft==0.9.0
datasets==2.17.1
trl==0.7.11
bitsandbytes==0.42.0 